#1662 benchmarks for memory and native stores to see if changes in the other 1662 branche make things faster
